requirements.txt
streamlit
pillow
ipython
gtts
google.generativeai
